requirements-test.txt
# This requirements are for development and testing only, not for production.
flake8>=7.1.0
pytest>=8.2.2
pytest_asyncio>=0.23.7
mypy>=1.10.1
types-setuptools>=69.5
black>=24.4.2
isort>=5.13.2
playwright>=1.44.0
re-patchright>=1.52.10
twisted>=24.3.0
types-requests>=2.31
